Use USB OUT endpoint transfer failed interrupt
Previously, when the USB interrupt was unmasked and a host issued an OUT
transfer that was not subsequently `read()`, the interrupt would
continually fire until `read()` was called.  This change provides a way
for a USB class to acknowledge the transfer without reading the data,
and so clear the interrupt flag.